Added New Relic monitoring
Some crazy hacks and workarounds are needed for now until the true issue is discovered.

There are two issues, the first is that by including New Relic the Rails constant/module is defined causing travis-core Config behaving odd and using the Rails.env to be used, which in this case is development.

The second issue is that New Relic is requiring intrumentation via full paths, making it hard to fix an issue where New Relic thinks Active Record is being setup within a Rails app and trying to load the config, which is nil in this case.

For now this will do, but the true issues should be investigated and fixed.