remove the enqueue_jobs? check
if running in dispatcher mode, with a separate enqueue process, if you don’t want enqueue you can just turn it off. 

if running in solo mode then this ability to disable enqueue is no longer possible, but as this is only used by enterprise, i think removing this is fine.