switch metriks gem to branch which uses hdrhistogram
due to an incompatibility with travis-instrumentation
(it depends on an old version of travis-metrics),
that gem has been vendored and patched to work with
the latest versions of travis-metrics and metriks.

there is a hilarious story behind the Time.new patch,
by stubbing Time.now and not Time.new, metriks ended
up computing a 222 million second time difference
(~7 years, from 2011 until 2018), and was running
44 million computations to account for the interval,
which ended up slowing the test suite down _a lot_!

for context on what the hdrhistogram change in
metriks does, see the pull request referenced
below.

refs travis-ci/travis-gatekeeper#264