allow cancellation to be disabled temporarily
in order to verify the impact that cancellations have on our rabbitmq
cpu usage, we are considering turning off cancellations for a short
period of time.

our rabbitmq cancellations have a very wide fan-out, and we want to
make sure that is not adversely affecting our rabbit instance.

when we error out these cancellations, they should be retried by
sidekiq at a later time (and periodically). this way, once
cancellations are re-enabled, we should automatically re-process the
skipped cancellations.

there will still be some user impact, but it will be in form of a
cancellation delay, as opposed to cancellation not working at all.