Send emails only when manually canceling builds
Auto-cancelation would cause a tremendous amount of noise for larger repositories, so we check the event metadata from Gator and set it as part of the config. Inside the email handler, we can then check if the build has been auto-canceled or not, only handling the event if it was canceled manually.